More networking improvement done
