+++ shared/src/commonMain/kotlin/com/dz/tmdbmobileapi/usecase/GetPopularMoviesUseCase.kt
@@ -1,7 +1,7 @@
 package com.dz.tmdbmobileapi.usecase
 
 import com.dz.tmdbmobileapi.base.Resource
-import com.dz.tmdbmobileapi.local.data.responses.MoviesResponse
+import com.dz.tmdbmobileapi.remote.responses.MoviesResponse
 import com.dz.tmdbmobileapi.remote.GetPopularMoviesDataSource
 
 class GetPopularMoviesUseCase(private val popularMoviesDataSource: GetPopularMoviesDataSource) {
